scripts/requirements.txt
boto3
botocore
PyYAML
ruamel.yaml
coverage
